Cast targets to float in binary label smoothing
Ensures targets are in floating point format before applying the smoothing formula in binary mode, preventing potential type mismatch errors during calculation.

+++ src/losses.py
@@ -31,7 +31,7 @@ class LabelSmoothingCrossEntropy(nn.Module):
         """
         if self.mode == "binary":
             logits = logits.squeeze(1)
-            targets = targets.squeeze(1)
+            targets = targets.squeeze(1).float()
 
         valid = targets != self.ignore_index
